feat: rename externalDb.js of the API at frontend
