+++ frontend/src/store/modules/idInModels.js
@@ -1,4 +1,4 @@
-import externalDbApi from '@/api/externalDb';
+import idInModelsApi from '@/api/idInModels';
 
 const data = {
   components: [],
@@ -7,7 +7,7 @@ const data = {
 
 const actions = {
   async getComponentsForExternalDb({ commit }, { dbName, externalId }) {
-    const { components, externalDb } = await externalDbApi.fetchComponentsForExternalDb({
+    const { components, externalDb } = await idInModelsApi.fetchComponentsForExternalDb({
       dbName,
       externalId,
     });
